lista las cards
