+++ js/script.js
@@ -1,3 +1,5 @@
+import { generarCard } from "./dinamic-template.js";
+
 function createLocalStorage() {
   //Genera un objeto que luego es agregado al local storage
   if (localStorage.length == 0) {
@@ -87,24 +89,39 @@ function createLocalStorage() {
 localStorage.clear();
 createLocalStorage();
 
-const $templateCard = document.getElementById("template-card").content,
-  $cards = document.querySelector(".cards"),
-  $fragment = document.createDocumentFragment();
+// export function generarCard(option) {
+//   const $templateCard = document.getElementById("template-card").content,
+//     $cards = document.querySelector(".cards"),
+//     $fragment = document.createDocumentFragment();
+//   //Genera tarjetas dinamicas en los html donde esta funcion sea llamada,dependiendo del option que le pase toma los datos del LocalSotorage
+//   let prueba = localStorage.getItem(option);
+//   prueba = JSON.parse(prueba);
+//   prueba.forEach((element, index) => {
+//     let { img1, img2, img3, alt, description, price } = element;
+//     $templateCard.querySelector("div").setAttribute("data-id", index);
+//     $templateCard.querySelector("img").setAttribute("src", img1);
+//     $templateCard.querySelector("img").setAttribute("alt", alt);
+//     $templateCard.querySelectorAll("p")[0].textContent = description;
+//     $templateCard.querySelectorAll("p")[1].textContent = `$${price}`;
+//     const $clon = document.importNode($templateCard, true);
+//     $fragment.appendChild($clon);
+//   });
 
-export function generarCard(option) {
-  //Genera tarjetas dinamicas en los html donde esta funcion sea llamada,dependiendo del option que le pase toma los datos del LocalSotorage
-  let prueba = localStorage.getItem(option);
-  prueba = JSON.parse(prueba);
-  prueba.forEach((element, index) => {
-    let { img1, img2, img3, alt, description, price } = element;
-    $templateCard.querySelector("div").setAttribute("data-id", index);
-    $templateCard.querySelector("img").setAttribute("src", img1);
-    $templateCard.querySelector("img").setAttribute("alt", alt);
-    $templateCard.querySelectorAll("p")[0].textContent = description;
-    $templateCard.querySelectorAll("p")[1].textContent = `$${price}`;
-    const $clon = document.importNode($templateCard, true);
-    $fragment.appendChild($clon);
-  });
+//   $cards.appendChild($fragment);
+// }
 
-  $cards.appendChild($fragment);
+console.log(document.location.pathname);
+
+let url = document.location.pathname;
+
+if (url.includes("male")) {
+  generarCard("hombre");
+}
+
+if (url.includes("female")) {
+  generarCard("mujer");
+}
+
+if (url.includes("show")) {
+  console.log("Lol");
 }
