+++ page/female.html
@@ -40,6 +40,6 @@
         </div>
       </section>
     </main>
-    <script src="../js/female.js" type="module"></script>
+    <script src="../js/script.js" type="module"></script>
   </body>
 </html>
