+++ page/male.html
@@ -39,6 +39,6 @@
         </div>
       </section>
     </main>
-    <script src="../js/male.js" type="module"></script>
+    <script src="../js/script.js" type="module"></script>
   </body>
 </html>
